Assignment 3/part 2/backend/requirements.txt
<<<<<<< HEAD
flask
pymongo
dnspython
=======
flask
pymongo
dnspython
>>>>>>> af418ba (Assignment 5)
python-dotenv